mmc: rockchip_sdhci: Rearrange and simplify regs and flags used
This rearrange and remove duplicate defines to make the code simpler.

There is no need to read vendor area1 and use an offset each time, it is
easier and clearer to just use the reg offset defined in TRM.

This also removes use of the misspelled CMDOUT reg, it will be re-added
when support for RK3588 is introduced.

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>