rockchip: mmc: rockchip_dw_mmc: Fix get_mmc_clk return value
The get_mmc_clk ops is expected to set a clock rate and return the
configured rate as an unsigned value. However, if clk_set_rate failes,
e.g. using a fixed rate clock, a negative error value is returned.

The mmc core will treat this as a valid unsigned rate and tries to
configure a divider based on this bogus clock rate.

Instead use 0 as the return value when setting clock rate fails to let
the mmc core be configured in bypass mode.

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>