gpio: rockchip: Use gpio alias id as gpio bank id
Use the gpio alias id as the gpio bank id, same as in the linux driver.

This reverts the recent change to base bank id on the gpio-ranges prop
and apply the same logic used when nodes was named gpio0@ to the alias.

With missing gpio aliases added to u-boot.dtsi files this should restore
gpio function for SoCs that have synced gpio0@ to gpio@ node name
changes from linux.

Fixes: 904b8700f81c ("gpio: gpio-rockchip: parse gpio-ranges for bank id")
Signed-off-by: Jonas Karlman <jonas@kwiboo.se>
Link: https://patchwork.ozlabs.org/project/uboot/patch/20230318235651.826148-3-jonas@kwiboo.se/