RFC: power: regulator: Only run autoset once for each regulator
Introduce a AUTOSET_DONE flag and use it to flag that autoset has run
for the regulator. Return -EALREADY on any subsequent call to
regulator_autoset.

FIXME: In this version autoset will only run once, even if first run
return an error. Unclear if this is the best aproach at this time.

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>